aes_gcm: Split x86-64 integrated implementation into its own module.
```
git difftool HEAD^1:src/aead/aes_gcm.rs \
                    src/aead/aes_gcm/aeshwclmulmovbe.rs
```
